refactor loads module
